Add files via upload
I worked in a team of four members to create a code that scraps data from the official NBA website (stats.nba.com/players), arranges it and presents it in form of easy to read statistics, plotted figures, and derived insights. The website contained lots of information such as statistics (per team, per player, etc.), history of every NBA team, information about current and retired players. The file is in the form of jupyter notebook and should be opened accordingly.